version 1.2 hisorique des actions et création de la fiche odl

+++ app/Http/Controllers/ActiviteController.php
@@ -3,13 +3,11 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Date;
 use Illuminate\Support\Facades\Auth;
-use Illuminate\Support\Facades\DB;
 use App\Models\Activite;
-use App\Models\Competence;
-use App\Models\Classe;
+use App\Models\Tache;
 use App\Models\User;
+use App\Models\Historique;
 
 class ActiviteController extends Controller
 {
@@ -22,21 +20,20 @@ class ActiviteController extends Controller
        *
        * @return \Illuminate\Http\Response
        */
-     // Afficher les activites appartenants a la personne qui s'est connecter
+     // Afficher les types utilisateurs
      public function index()
      {
         $this->authorize('ad_re_su', User::class);
        try
        {
-         $activites = Activite::select('*')->get();
+            $activites = Activite::select('*')->orderBy('id','DESC')->get();
 
-         return view('activites.index', compact('activites'));
+            return view('activites.index', compact('activites'));
+        }
+        catch(\Exception $exception)
+       {
+           return redirect('erreur')->with('messageerreur',$exception->getMessage());
        }
-      catch(\Exception $exception)
-      {
-          return redirect('erreur')->with('messageerreur',$exception->getMessage());
-      }
-
      }
 
        /**
@@ -50,19 +47,15 @@ class ActiviteController extends Controller
         $this->authorize('ad_re_su', User::class);
        try
        {
-         $activite = new Activite();
-         $competences = Competence::select('*')->get();
-         $classes = Classe::select('*')->where('libelleclasse','not like',"%Aucune%")->get();
+          $activite = new Activite();
 
-         return view('activites.create',compact('activite','competences','classes'));
-
+          return view('activites.create',compact('activite'));
         }
         catch(\Exception $exception)
-        {
-            return redirect('erreur')->with('messageerreur',$exception->getMessage());
-        }
+       {
+           return redirect('erreur')->with('messageerreur',$exception->getMessage());
+       }
      }
-
 
        /**
       * Store a newly created resource in storage.
@@ -74,20 +67,19 @@ class ActiviteController extends Controller
      public function store()
      {
         $this->authorize('ad_re_su', User::class);
-        $this->validator();
-
        try
        {
+        if(Activite::where('libelleactivite','=',request('libelleactivite'))->select('id')->doesntExist())
+        {
+            $libelle = request('libelleactivite');
 
-         $activite = Activite::create([
-          'libelleactivite'=> request('libelleactivite'),
-          'identifiantactivite'=> request('identifiantactivite'),
-          'competence_id'=> request('competence_id'),
-          'classe_id'=> request('classe_id'),
-          ]);
+            $activite = Activite::create($this->validator());
 
-          return redirect('activites/create')->with('message', 'Informations bien enregistrées.');
+            $this->historique(request('libelleactivite'), 'Ajout');
 
+            return redirect('activites/create')->with('message', 'Activité bien ajoutée.');
+        }
+        return back()->with('messagealert',"Cette activité existe déjà.");
       }
       catch(\Exception $exception)
       {
@@ -95,7 +87,6 @@ class ActiviteController extends Controller
       }
      }
 
-     // return redirect('rapports')->with('message', 'Rapport bien ajoutée.');
       /**
       * Display the specified resource.
       *
@@ -106,14 +97,14 @@ class ActiviteController extends Controller
      public function show(Activite $activite)
      {
         $this->authorize('ad_re_su', User::class);
-        try
-        {
-           return view('activites.show',compact('activite'));
+       try
+       {
+          return view('activites.show',compact('activite'));
         }
         catch(\Exception $exception)
-        {
-            return redirect('erreur')->with('messageerreur',$exception->getMessage());
-        }
+       {
+           return redirect('erreur')->with('messageerreur',$exception->getMessage());
+       }
      }
 
     /**
@@ -128,15 +119,12 @@ class ActiviteController extends Controller
         $this->authorize('ad_re_su', User::class);
        try
        {
-        $competences = Competence::all();
-        $classes = Classe::all();
-
-        return view('activites.edit', compact('activite','competences','classes'));
+          return view('activites.edit', compact('activite'));
+        }
+        catch(\Exception $exception)
+       {
+           return redirect('erreur')->with('messageerreur',$exception->getMessage());
        }
-      catch(\Exception $exception)
-      {
-          return redirect('erreur')->with('messageerreur',$exception->getMessage());
-      }
      }
 
         /**
@@ -150,24 +138,22 @@ class ActiviteController extends Controller
      public function update(Activite $activite)
      {
         $this->authorize('ad_re_su', User::class);
-      try
-      {
-              /** actualisation de l'activite **/
-              $activite->update([
-                'libelleactivite'=> request('libelleactivite'),
-                'identifiantactivite'=> request('identifiantactivite'),
-                'competence_id'=> request('competence_id'),
-                'classe_id'=> request('classe_id'),
-                ]);
+       try
+       {
+          $activite->update([
+              'libelleactivite'=> request('libelleactivite'),
+              'identifiantactivite'=> request('identifiantactivite'),
+              'categorie'=> request('categorie'),
+          ]);
 
-            return redirect('activites/' . $activite->id)->with('message', 'Activité bien mise à jour.');
+          $this->historique(request('libelleactivite'), 'Modification');
 
+          return redirect('activites/' . $activite->id);
         }
         catch(\Exception $exception)
-        {
-            return redirect('erreur')->with('messageerreur',$exception->getMessage());
-        }
-
+       {
+           return redirect('erreur')->with('messageerreur',$exception->getMessage());
+       }
      }
 
       /**
@@ -177,30 +163,49 @@ class ActiviteController extends Controller
       * @return \Illuminate\Http\Response
       */
 
-     public function destroy(Activite $activite)
+     public function destroy(activite $activite)
      {
         $this->authorize('ad_re_su', User::class);
        try
        {
+        if(Tache::where('activite_id','=',$activite->id)->select('id')->exists())
+        {
+           return back()->with('messagealert',"Suppression pas possible. Cette activité est référencée dans une autre table.");
+        }
+        else
+        {
             $activite->delete();
 
+            $this->historique($activite->libelleactivite, 'Suppression');
+
             return redirect('activites')->with('messagealert','Suppression éffectuée');
-      }
-      catch(\Exception $exception)
-      {
-          return redirect('erreur')->with('messageerreur',$exception->getMessage());
-      }
-
+        }
+       }
+        catch(\Exception $exception)
+       {
+           return redirect('erreur')->with('messageerreur',$exception->getMessage());
+       }
      }
-
 
      private  function validator()
      {
          return request()->validate([
-            'libelleactivite'=> 'required|min:2',
-            'identifiantactivite'=> 'required|max:10',
-            'competence_id' =>'required|integer',
-            'classe_id' =>'required|integer',
+             'libelleactivite'=>'required|min:2',
+             'identifiantactivite'=>'max:10',
+             'categorie'=>'max:255',
          ]);
      }
+
+     private function historique($attribute, $action)
+    {
+        $auth_user = (Auth::user()->nomuser). ' ' .(Auth::user()->prenomuser);
+
+        /** historiques des actions sur le systeme **/
+        $historique = Historique::create([
+        'user_action'=> $auth_user,
+        'table'=> 'Activite',
+        'attribute' => $attribute,
+        'action'=> $action
+        ]);
+    }
 }
